+++ lib/src/result.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 class Result<T, E> {
   final bool _ok;
   final T? _artifact;
@@ -38,6 +40,52 @@ class Result<T, E> {
       throw Exception("Called unwrapErr() on an Ok value");
     }
     return err();
+  }
+
+  T when({
+    required T Function(T artifact) ok,
+    required T Function(E error) err,
+  }) {
+    return isOk() ? ok(_artifact!) : err(_error!);
+  }
+
+  FutureOr<R> fold<R>({
+    required FutureOr<R> Function(T artifact) ok,
+    required FutureOr<R> Function(E error) err,
+  }) async {
+    return isOk() ? await ok(_artifact!) : await err(_error!);
+  }
+
+  Result<RT, E> map<RT>(RT Function(T artifact) mapper) {
+    return isOk() ? Ok<RT, E>(mapper(_artifact!)) : Err<RT, E>(_error!);
+  }
+  Result<T, RE> mapErr<RE>(RE Function(E error) mapper) {
+    return isOk() ? Ok<T, RE>(_artifact!) : Err<T, RE>(mapper(_error!));
+  }
+
+  RT mapOr<RT>(RT Function(T artifact) mapper, RT defaultValue) {
+    return isOk() ? mapper(_artifact!) : defaultValue;
+  }
+  RT mapOrElse<RT>(RT Function(T artifact) mapper, RT Function(E error) errMapper) {
+    return isOk() ? mapper(_artifact!) : errMapper(_error!);
+  }
+  RE mapOrErr<RE>(RE Function(E error) mapper, RE defaultValue) {
+    return isOk() ? defaultValue : mapper(_error!);
+  }
+
+  RT mapErrOr<RT>(RT Function(E error) mapper, RT defaultValue) {
+    return isOk() ? defaultValue : mapper(_error!);
+  }
+
+  T mapErrOrElse<T>(T Function(E error) mapper, T defaultValue) {
+    return isOk() ? defaultValue : mapper(_error!);
+  }
+
+  T unwrapOr(T defaultValue) {
+    return isOk() ? _artifact! : defaultValue;
+  }
+  T unwrapOrElse(T Function(E error) defaultValue) {
+    return isOk() ? _artifact! : defaultValue(_error!);
   }
 }